+++ src/main/java/com/hutchison/calendar/intcode/operation/bool/BooleanOperation.java
@@ -13,15 +13,15 @@ import static com.hutchison.calendar.intcode.operation.ParamMode.fromChar;
 
 public abstract class BooleanOperation implements Operation {
 
-    protected final Codes storeBoolean(Codes incomingCodes, BiPredicate<Integer, Integer> predicate) {
-        List<Integer> codes = incomingCodes.getCodes();
+    protected final Codes storeBoolean(Codes incomingCodes, BiPredicate<Double, Double> predicate) {
+        List<Double> codes = incomingCodes.getCodes();
         int cursor = incomingCodes.getCursor();
         validateCursorPosition(cursor + 3, codes.size());
         String opString = getOpCodeString(codes.get(cursor));
-        int val1 = getValueFromCodes(codes, fromChar(opString.charAt(0)), cursor + 1);
-        int val2 = getValueFromCodes(codes, fromChar(opString.charAt(1)), cursor + 2);
-        int valToStore = predicate.test(val1, val2) ? 1 : 0;
-        codes.set(codes.get(cursor + 3), valToStore);
+        double val1 = getValueFromCodes(codes, fromChar(opString.charAt(0)), cursor + 1);
+        double val2 = getValueFromCodes(codes, fromChar(opString.charAt(1)), cursor + 2);
+        double valToStore = predicate.test(val1, val2) ? 1 : 0;
+        codes.set(codes.get(cursor + 3).intValue(), valToStore);
 
         return incomingCodes.toBuilder()
                 .codes(codes)
